requirements.txt
python-telegram-bot==20.7
torch
opencv-python
numpy
Pillow
gdown